Create Rounding the array of real numbers to the nearest integer and checking for repeatability
Enter an array of real numbers from the keyboard (the maximum dimension is 45 elements). Create a new integer array to write to
the values ​​of the elements of the entered array, rounded to the nearest whole numbers. Ensure that no number is repeated in the new array. Print the generated array.