Create enter the matrix with dimension r*s. write down the upper values ​​operation for the last row of the XOR matrix
Allocate space for the matrix with dimension R  S (declare R and S as macro-constants or constants) whose elements are single-byte numbers. Fill R-1 rows of the matrix with random numbers having a value of 0 or 1. The formed rows of the matrix are binary integer codes of bits S. The last R-th line of the matrix must contain the result of XOR operations on all R-1 generated binary codes. Result calculate bit by bit without converting numbers to decimals. Print the received matrix.